server/requirements.txt
asgiref==3.5.2
bcrypt==4.0.1
certifi==2022.6.15
click==8.1.3
ecdsa==0.18.0
fastapi==0.68.2
h11==0.13.0
networkx==2.8.8
passlib==1.7.4
psycopg2==2.8.6
pyasn1==0.4.8
pydantic==1.9.2
python-jose==3.3.0
python-multipart==0.0.5
requests==2.28.1
rsa==4.9
six==1.16.0
starlette==0.14.2
typing_extensions==4.3.0
uvicorn==0.15.0
